api/requirements.txt
openai==1.14.0
httpx==0.27.2
python-dotenv==1.0.1
flask==3.0.0
requests==2.31.0
python-json-logger==2.0.7 